Added 'chars' (size) option and added 'multi' and header.
Multi is a new type that allows for aggregated form elements. Such as
'[first name] [last name] on one line.

A sizing paramater was added. 'size' already has meaning in liquid, so
the var name 'chars' was chosen.

@pilz Lots of new goodies. More and more complex forms are possible now
with the form generator. Would you know who could help to start
documenting this?